[crunchyroll] preparations for re-write
